КОММЕНТ

+++ lib/features/weather/domain/entities/weather_entity.dart
@@ -79,7 +79,7 @@ class RainOrSnowEntity {
 class WindEntity {
   final double speed;
   final int deg;
-  final int? gust;
+  final double? gust;
   WindEntity({required this.speed, required this.deg, required this.gust});
 }
